+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Honeybadger integration
+
 ## 0.1.0
 
 - First release